gears/cyberinstaller.py: Move CyberInstaller to its own module.
